Refac: 클래스 파일 분리, 접근한정자 추가
