+++ 후/src/mission2_0/RandomGacha.java
@@ -3,35 +3,6 @@ package mission2_0;
 import java.util.Random;
 import java.util.Scanner;
 
-class CocoaMember {
-
-    private String name;
-
-    CocoaMember(String name) {
-        this.name = name;
-    }
-
-    String getName() {
-        return name;
-    }
-
-    static CocoaMember[] makeSquad1() {
-        return new CocoaMember[]{
-            new CocoaMember("Donggi"),
-            new CocoaMember("K"),
-            new CocoaMember("Tany"),
-            new CocoaMember("Taksu"),
-            new CocoaMember("Hoo"),
-            new CocoaMember("Jerry"),
-            new CocoaMember("Nohri"),
-            new CocoaMember("MK"),
-            new CocoaMember("ttasjwi"),
-            new CocoaMember("PO"),
-            new CocoaMember("Mandoo")
-        };
-    }
-}
-
 class RandomGacha {
 
     private int coin = 0;
@@ -96,9 +67,6 @@ class RandomGacha {
         }
         System.out.println("...끝!");
     }
-}
-
-class RandomGachaTest {
 
     public static void main(String[] args) {
         RandomGacha gacha = new RandomGacha();
